refactor: move contest types to src/lib/contests/types/
